fix f32 to f64

+++ src/opt.rs
@@ -24,7 +24,7 @@ pub struct TspNode {
     opt: Option<Vec<usize>>,
     #[get = "pub"]
     #[set = "pub"]
-    bias: f32,
+    bias: f64,
     max_dist: Option<f64>,
 }
 
@@ -234,7 +234,7 @@ impl From<TspNode> for QuboModel {
                         let jb = j + tsp.dim * b;
                         if i == j {
                             Q[[i + tsp.dim * a, ((i + 1) % tsp.dim) + tsp.dim * b]] +=
-                                tsp.distance(a, b) as f32;
+                                tsp.distance(a, b);
                             Q[[ia, jb]] += tsp.bias;
                         }
